Resolve 78, Resolve 91, refactor server
This change was specifically to resolve issue 78. During the course of the work it became apparent is was a good opportunity to provide some additional organization to the server and resolve issue #91 at the same time.
